Rename tools -> test and create two directories inside generic and static
