+++ MANIFEST.in
@@ -8,7 +8,7 @@ include requirements.txt
 recursive-include moduleframework *
 recursive-include docs *
 recursive-include examples *
-recursive-include tools *
+recursive-include tests *
 recursive-include distro *
 recursive-include man *
 recursive-include build_manpages *.py
