+++ moduleframework/common.py
@@ -91,7 +91,8 @@ DEFAULTNSPAWNTIMEOUT = 10
 MODULE_DEFAULT_PROFILE = "default"
 TRUE_VALUES_DICT = ['yes', 'YES', 'yes', 'True', 'true', 'ok', 'OK']
 OPENSHIFT_INIT_WAIT = 50
-STATIC_CHECKS = ['static_dockerfiule_checks', 'helpmd_lint']
+STATIC_LINTERS = 'static'
+GENERIC_TEST = 'generic'
 
 def generate_unique_name(size=10):
     return ''.join(random.choice(string.ascii_lowercase) for _ in range(size))
