+++ moduleframework/tests/generic/check_compose.py
